Architecture Pattern Improved.
