+++ zhttp/src/main/java/com/zkrallah/zhttp/core/ZGet.kt
@@ -1,8 +1,14 @@
-package com.zkrallah.zhttp
+package com.zkrallah.zhttp.core
 
 import android.util.Log
 import com.google.gson.JsonParseException
-import com.zkrallah.zhttp.Helper.deserializeBody
+import com.zkrallah.zhttp.model.Header
+import com.zkrallah.zhttp.util.Helper.deserializeBody
+import com.zkrallah.zhttp.model.HttpResponse
+import com.zkrallah.zhttp.model.Query
+import com.zkrallah.zhttp.model.Response
+import com.zkrallah.zhttp.util.UrlEncoderUtil
+import com.zkrallah.zhttp.client.ZHttpClient
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Deferred
 import kotlinx.coroutines.Dispatchers
@@ -17,25 +23,24 @@ import java.net.SocketTimeoutException
 import java.net.URL
 
 /**
- * ZPut class for handling PUT HTTP requests.
+ * ZGet class for handling GET HTTP requests.
  *
  * @param client ZHttpClient instance of the client.
  */
-class ZPut(val client: ZHttpClient) {
+class ZGet(val client: ZHttpClient) {
 
     /**
-     * Executes a raw PUT HTTP request synchronously.
+     * Executes a raw GET HTTP request synchronously.
      *
      * @param endpoint Endpoint to append to the base URL.
-     * @param requestBody Body of the request.
      * @param queries List of query parameters to include in the URL.
      * @param headers List of headers to include in the request.
      * @return HttpResponse containing the response details.
      */
     @Synchronized
     @Throws(Exception::class)
-    fun doPut(
-        endpoint: String, requestBody: Any, queries: List<Query>?, headers: List<Header>?
+    fun doGet(
+        endpoint: String, queries: List<Query>?, headers: List<Header>?
     ): HttpResponse? {
         // Build the full URL with endpoint and query parameters
         val urlString = StringBuilder(client.getBaseUrl()).append("/").append(endpoint)
@@ -47,9 +52,8 @@ class ZPut(val client: ZHttpClient) {
         connection.readTimeout = client.getReadTimeout()
 
         return try {
-            // Set the request method to PUT
-            connection.requestMethod = PUT
-            connection.doOutput = true
+            // Set the request method to GET
+            connection.requestMethod = GET
 
             // Add default headers from the ZHttpClient
             client.getDefaultHeaders().forEach { (key, value) ->
@@ -59,15 +63,6 @@ class ZPut(val client: ZHttpClient) {
             // Add headers to the request
             headers?.forEach { (key, value) ->
                 connection.addRequestProperty(key, value)
-            }
-
-            // Serialize the request body to JSON
-            val body = client.getGsonInstance().toJson(requestBody)
-
-            // Write the request body to the output stream
-            connection.outputStream.use { outputStream ->
-                outputStream.write(body.toByteArray())
-                outputStream.flush()
             }
 
             val response = StringBuilder()
@@ -80,11 +75,11 @@ class ZPut(val client: ZHttpClient) {
                 }
             } catch (e: SocketTimeoutException) {
                 // If a socket timeout occurs, return an HttpResponse with the exception
-                Log.e(TAG, "doPut: $e", e)
+                Log.e(TAG, "doGet: $e", e)
                 return HttpResponse(exception = e)
             } catch (e: Exception) {
                 // If there's an error, read the error stream for additional information
-                Log.e(TAG, "doPut: $e", e)
+                Log.e(TAG, "doGet: $e", e)
                 BufferedReader(InputStreamReader(connection.errorStream)).use { reader ->
                     var line: String?
                     while (reader.readLine().also { line = it } != null) response.append(line)
@@ -97,11 +92,11 @@ class ZPut(val client: ZHttpClient) {
                 response.toString(),
                 connection.headerFields,
                 connection.date,
-                connection.permission
+                connection.permission,
             )
         } catch (e: Exception) {
             // If an exception occurs, log the error and return an HttpResponse with the exception
-            Log.e(TAG, "doPut: $e", e)
+            Log.e(TAG, "doGet: $e", e)
             HttpResponse(exception = e)
         } finally {
             // Disconnect the connection when done
@@ -110,23 +105,22 @@ class ZPut(val client: ZHttpClient) {
     }
 
     /**
-     * Performs a suspended PUT HTTP request asynchronously, returning a [Deferred] object containing the result.
+     * Performs a suspended GET HTTP request asynchronously, returning a [Deferred] object containing the result.
      *
-     * @param endpoint The endpoint URL to send the PUT request to.
-     * @param requestBody The request body to include in the PUT request.
+     * @param endpoint The endpoint URL to send the GET request to.
      * @param queries The list of query parameters to include in the request.
      * @param headers The list of headers to include in the request.
-     * @return A [Deferred] object containing the result of the PUT request.
+     * @return A [Deferred] object containing the result of the GET request.
      */
-    suspend fun doSuspendedPutRequest(
-        endpoint: String, requestBody: Any, queries: List<Query>?, headers: List<Header>?
+    suspend fun doSuspendedGetRequest(
+        endpoint: String, queries: List<Query>?, headers: List<Header>?
     ): Deferred<HttpResponse?> {
         return withContext(Dispatchers.IO) {
             async {
                 try {
-                    doPut(endpoint, requestBody, queries, headers)
+                    doGet(endpoint, queries, headers)
                 } catch (e: Exception) {
-                    Log.e(TAG, "doSuspendedPutRequest: $e", e)
+                    Log.e(TAG, "doSuspendedGetRequest: $e", e)
                     val response = HttpResponse(exception = e)
                     response
                 }
@@ -135,22 +129,20 @@ class ZPut(val client: ZHttpClient) {
     }
 
     /**
-     * Processes a PUT HTTP request asynchronously.
+     * Processes a GET HTTP request asynchronously.
      *
-     * @param endpoint The endpoint URL to send the PUT request to.
-     * @param requestBody The request body to include in the PUT request.
+     * @param endpoint The endpoint URL to send the GET request to.
      * @param queries The list of query parameters to include in the request.
      * @param headers The list of headers to include in the request.
-     * @return A [Response] object containing the result of the PUT request, or `null` if an error occurs.
+     * @return A [Response] object containing the result of the GET request, or `null` if an error occurs.
      */
-    suspend inline fun <reified T> processPut(
-        endpoint: String, requestBody: Any, queries: List<Query>?, headers: List<Header>?
+    suspend inline fun <reified T> processGet(
+        endpoint: String, queries: List<Query>?, headers: List<Header>?
     ): Response<T>? {
-        val response =
-            doSuspendedPutRequest(endpoint, requestBody, queries, headers).await() ?: return null
+        val response = doSuspendedGetRequest(endpoint, queries, headers).await() ?: return null
 
         response.exception?.let {
-            Log.e("ZPut", "processPut: $it", it)
+            Log.e("ZGet", "processGet: $it", it)
         }
 
         val body = client.getGsonInstance().deserializeBody<T>(response.body)
@@ -167,24 +159,22 @@ class ZPut(val client: ZHttpClient) {
     }
 
     /**
-     * Process a PUT HTTP request with callback for the response.
+     * Process a GET HTTP request with callback for the response.
      *
      * @param endpoint Endpoint to append to the base URL.
-     * @param requestBody Body of the request.
-     * @param queries List of query parameters to include in the URL.
      * @param headers List of headers to include in the request.
+     * @param queries List of query parameters to include in the URL.
      * @param onComplete Callback to handle the HttpResponse or an exception.
      * @return Job that will be completed with the HttpResponse or an exception.
      */
-    inline fun <reified T> processPut(
+    inline fun <reified T> processGet(
         endpoint: String,
-        requestBody: Any,
         queries: List<Query>?,
         headers: List<Header>?,
         crossinline onComplete: (success: Response<T>?, failure: Exception?) -> Unit
     ): Job {
         return CoroutineScope(Dispatchers.IO).launch {
-            val response = doSuspendedPutRequest(endpoint, requestBody, queries, headers).await()
+            val response = doSuspendedGetRequest(endpoint, queries, headers).await()
 
             if (response?.code == null) {
                 onComplete(null, NullPointerException("Could not make request."))
@@ -211,7 +201,8 @@ class ZPut(val client: ZHttpClient) {
     }
 
     companion object {
-        private const val TAG = "ZPut"
-        private const val PUT = "PUT"
+        private const val TAG = "ZGet"
+        private const val GET = "GET"
     }
+
 }
