+++ zhttp/src/main/java/com/zkrallah/zhttp/core/ZPatch.kt
@@ -1,8 +1,14 @@
-package com.zkrallah.zhttp
+package com.zkrallah.zhttp.core
 
 import android.util.Log
 import com.google.gson.JsonParseException
-import com.zkrallah.zhttp.Helper.deserializeBody
+import com.zkrallah.zhttp.util.Helper.deserializeBody
+import com.zkrallah.zhttp.client.ZHttpClient
+import com.zkrallah.zhttp.model.Header
+import com.zkrallah.zhttp.model.HttpResponse
+import com.zkrallah.zhttp.model.Query
+import com.zkrallah.zhttp.model.Response
+import com.zkrallah.zhttp.util.UrlEncoderUtil
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Deferred
 import kotlinx.coroutines.Dispatchers
@@ -11,33 +17,31 @@ import kotlinx.coroutines.async
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import java.io.BufferedReader
-import java.io.DataOutputStream
-import java.io.FileInputStream
 import java.io.InputStreamReader
 import java.net.HttpURLConnection
 import java.net.SocketTimeoutException
 import java.net.URL
 
 /**
- * ZMultipart class for handling multipart/form-data HTTP requests.
+ * ZPatch class for handling PATCH HTTP requests.
  *
  * @param client ZHttpClient instance of the client.
  */
-class ZMultipart(val client: ZHttpClient) {
+class ZPatch(val client: ZHttpClient) {
 
     /**
-     * Executes a raw multipart/form-data POST HTTP request synchronously.
+     * Executes a raw PATCH HTTP request synchronously.
      *
      * @param endpoint Endpoint to append to the base URL.
-     * @param parts List of MultipartBody parts to include in the request.
+     * @param requestBody Body of the request.
      * @param queries List of query parameters to include in the URL.
      * @param headers List of headers to include in the request.
      * @return HttpResponse containing the response details.
      */
     @Synchronized
     @Throws(Exception::class)
-    fun doMultipart(
-        endpoint: String, parts: List<MultipartBody>, queries: List<Query>?, headers: List<Header>?
+    fun doPatch(
+        endpoint: String, requestBody: Any, queries: List<Query>?, headers: List<Header>?
     ): HttpResponse? {
         // Build the full URL with endpoint and query parameters
         val urlString = StringBuilder(client.getBaseUrl()).append("/").append(endpoint)
@@ -49,8 +53,8 @@ class ZMultipart(val client: ZHttpClient) {
         connection.readTimeout = client.getReadTimeout()
 
         return try {
-            // Set the request method to POST
-            connection.requestMethod = POST
+            // Set the request method to PATCH
+            connection.requestMethod = PATCH
             connection.doOutput = true
 
             // Add default headers from the ZHttpClient
@@ -59,44 +63,18 @@ class ZMultipart(val client: ZHttpClient) {
             }
 
             // Add headers to the request
-            connection.setRequestProperty("Content-Type", "multipart/form-data; boundary=$BOUNDARY")
             headers?.forEach { (key, value) ->
                 connection.addRequestProperty(key, value)
             }
 
-            // Opens the DataOutputStream to write the request body
-            val outputStream = DataOutputStream(connection.outputStream)
+            // Serialize the request body to JSON
+            val body = client.getGsonInstance().toJson(requestBody)
 
-            // Handle whether the multipart body is a file or not
-            parts.forEach { part ->
-                // Declares the beginning of a multipart
-                outputStream.writeBytes("--$BOUNDARY\r\n")
-
-                // If the multipart is not a file
-                part.body?.let { requestBody ->
-                    outputStream.writeBytes("Content-Disposition: form-data; name=\"${part.name}\"\r\n")
-                    part.contentType?.let { outputStream.writeBytes("Content-Type: $it\r\n") }
-                    outputStream.writeBytes("\r\n")
-
-                    outputStream.write(requestBody.toByteArray())
-                }
-
-                // If the multipart is a file
-                part.filePath?.let { filePath ->
-                    outputStream.writeBytes("Content-Disposition: form-data; name=\"${part.fileName}\"; filename=\"$filePath\"\r\n")
-                    part.contentType?.let { outputStream.writeBytes("Content-Type: $it\r\n") }
-                    outputStream.writeBytes("\r\n")
-
-                    FileInputStream(filePath).use { fileInputStream ->
-                        fileInputStream.copyTo(outputStream, bufferSize = client.getBufferSize())
-                    }
-                }
-                outputStream.writeBytes("\r\n")
+            // Write the request body to the output stream
+            connection.outputStream.use { outputStream ->
+                outputStream.write(body.toByteArray())
+                outputStream.flush()
             }
-
-            // Declares the end of the request body
-            outputStream.writeBytes("\r\n--$BOUNDARY--\r\n")
-            outputStream.flush()
 
             val response = StringBuilder()
 
@@ -108,11 +86,11 @@ class ZMultipart(val client: ZHttpClient) {
                 }
             } catch (e: SocketTimeoutException) {
                 // If a socket timeout occurs, return an HttpResponse with the exception
-                Log.e(TAG, "doMultipart: $e", e)
+                Log.e(TAG, "doPatch: $e", e)
                 return HttpResponse(exception = e)
             } catch (e: Exception) {
                 // If there's an error, read the error stream for additional information
-                Log.e(TAG, "doMultipart: $e", e)
+                Log.e(TAG, "doPatch: $e", e)
                 BufferedReader(InputStreamReader(connection.errorStream)).use { reader ->
                     var line: String?
                     while (reader.readLine().also { line = it } != null) response.append(line)
@@ -129,7 +107,7 @@ class ZMultipart(val client: ZHttpClient) {
             )
         } catch (e: Exception) {
             // If an exception occurs, log the error and return an HttpResponse with the exception
-            Log.e(TAG, "doMultipart: $e", e)
+            Log.e(TAG, "doPatch: $e", e)
             HttpResponse(exception = e)
         } finally {
             // Disconnect the connection when done
@@ -138,23 +116,23 @@ class ZMultipart(val client: ZHttpClient) {
     }
 
     /**
-     * Performs a suspended multipart HTTP request asynchronously, returning a [Deferred] object containing the result.
+     * Performs a suspended PATCH HTTP request asynchronously, returning a [Deferred] object containing the result.
      *
-     * @param endpoint The endpoint URL to send the multipart request to.
-     * @param parts The list of multipart parts to include in the request.
+     * @param endpoint The endpoint URL to send the PATCH request to.
+     * @param requestBody The request body to include in the PATCH request.
      * @param queries The list of query parameters to include in the request.
      * @param headers The list of headers to include in the request.
-     * @return A [Deferred] object containing the result of the multipart request.
+     * @return A [Deferred] object containing the result of the PATCH request.
      */
-    suspend fun doSuspendedMultipartRequest(
-        endpoint: String, parts: List<MultipartBody>, queries: List<Query>?, headers: List<Header>?
+    suspend fun doSuspendedPatchRequest(
+        endpoint: String, requestBody: Any, queries: List<Query>?, headers: List<Header>?
     ): Deferred<HttpResponse?> {
         return withContext(Dispatchers.IO) {
             async {
                 try {
-                    doMultipart(endpoint, parts, queries, headers)
+                    doPatch(endpoint, requestBody, queries, headers)
                 } catch (e: Exception) {
-                    Log.e(TAG, "doSuspendedMultipartRequest: $e", e)
+                    Log.e(TAG, "doSuspendedPatchRequest: $e", e)
                     val response = HttpResponse(exception = e)
                     response
                 }
@@ -163,22 +141,22 @@ class ZMultipart(val client: ZHttpClient) {
     }
 
     /**
-     * Processes a multipart HTTP request asynchronously.
+     * Processes a PATCH HTTP request asynchronously.
      *
-     * @param endpoint The endpoint URL to send the multipart request to.
-     * @param parts The list of multipart parts to include in the request.
+     * @param endpoint The endpoint URL to send the PATCH request to.
+     * @param requestBody The request body to include in the PATCH request.
      * @param queries The list of query parameters to include in the request.
      * @param headers The list of headers to include in the request.
-     * @return A [Response] object containing the result of the multipart request, or `null` if an error occurs.
+     * @return A [Response] object containing the result of the PATCH request, or `null` if an error occurs.
      */
-    suspend inline fun <reified T> processMultiPart(
-        endpoint: String, parts: List<MultipartBody>, queries: List<Query>?, headers: List<Header>?
+    suspend inline fun <reified T> processPatch(
+        endpoint: String, requestBody: Any, queries: List<Query>?, headers: List<Header>?
     ): Response<T>? {
         val response =
-            doSuspendedMultipartRequest(endpoint, parts, queries, headers).await() ?: return null
+            doSuspendedPatchRequest(endpoint, requestBody, queries, headers).await() ?: return null
 
         response.exception?.let {
-            Log.e("ZMultipart", "processMultiPart: $it", it)
+            Log.e("ZPatch", "processPatch: $it", it)
         }
 
         val body = client.getGsonInstance().deserializeBody<T>(response.body)
@@ -195,24 +173,24 @@ class ZMultipart(val client: ZHttpClient) {
     }
 
     /**
-     * Process a multipart/form-data POST HTTP request with callback for the response.
+     * Process a PATCH HTTP request with callback for the response.
      *
      * @param endpoint Endpoint to append to the base URL.
-     * @param parts List of MultipartBody parts to include in the request.
+     * @param requestBody Body of the request.
      * @param queries List of query parameters to include in the URL.
      * @param headers List of headers to include in the request.
      * @param onComplete Callback to handle the HttpResponse or an exception.
      * @return Job that will be completed with the HttpResponse or an exception.
      */
-    inline fun <reified T> processMultiPart(
+    inline fun <reified T> processPatch(
         endpoint: String,
-        parts: List<MultipartBody>,
-        headers: List<Header>?,
+        requestBody: Any,
         queries: List<Query>?,
+        headers: List<Header>?,
         crossinline onComplete: (success: Response<T>?, failure: Exception?) -> Unit
     ): Job {
         return CoroutineScope(Dispatchers.IO).launch {
-            val response = doSuspendedMultipartRequest(endpoint, parts, queries, headers).await()
+            val response = doSuspendedPatchRequest(endpoint, requestBody, queries, headers).await()
 
             if (response?.code == null) {
                 onComplete(null, NullPointerException("Could not make request."))
@@ -239,8 +217,7 @@ class ZMultipart(val client: ZHttpClient) {
     }
 
     companion object {
-        private const val TAG = "ZMultipart"
-        private const val POST = "POST"
-        private const val BOUNDARY = "*****"
+        private const val TAG = "ZPatch"
+        private const val PATCH = "PATCH"
     }
 }
