+++ zhttp/src/main/java/com/zkrallah/zhttp/core/ZPost.kt
@@ -1,8 +1,14 @@
-package com.zkrallah.zhttp
+package com.zkrallah.zhttp.core
 
 import android.util.Log
 import com.google.gson.JsonParseException
-import com.zkrallah.zhttp.Helper.deserializeBody
+import com.zkrallah.zhttp.util.Helper.deserializeBody
+import com.zkrallah.zhttp.client.ZHttpClient
+import com.zkrallah.zhttp.model.Header
+import com.zkrallah.zhttp.model.HttpResponse
+import com.zkrallah.zhttp.model.Query
+import com.zkrallah.zhttp.model.Response
+import com.zkrallah.zhttp.util.UrlEncoderUtil
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Deferred
 import kotlinx.coroutines.Dispatchers
@@ -17,24 +23,25 @@ import java.net.SocketTimeoutException
 import java.net.URL
 
 /**
- * ZGet class for handling GET HTTP requests.
+ * ZPost class for handling POST HTTP requests.
  *
  * @param client ZHttpClient instance of the client.
  */
-class ZGet(val client: ZHttpClient) {
+class ZPost(val client: ZHttpClient) {
 
     /**
-     * Executes a raw GET HTTP request synchronously.
+     * Executes a raw POST HTTP request synchronously.
      *
      * @param endpoint Endpoint to append to the base URL.
+     * @param requestBody Body of the request.
      * @param queries List of query parameters to include in the URL.
      * @param headers List of headers to include in the request.
      * @return HttpResponse containing the response details.
      */
     @Synchronized
     @Throws(Exception::class)
-    fun doGet(
-        endpoint: String, queries: List<Query>?, headers: List<Header>?
+    fun doPost(
+        endpoint: String, requestBody: Any, queries: List<Query>?, headers: List<Header>?
     ): HttpResponse? {
         // Build the full URL with endpoint and query parameters
         val urlString = StringBuilder(client.getBaseUrl()).append("/").append(endpoint)
@@ -46,8 +53,9 @@ class ZGet(val client: ZHttpClient) {
         connection.readTimeout = client.getReadTimeout()
 
         return try {
-            // Set the request method to GET
-            connection.requestMethod = GET
+            // Set the request method to POST
+            connection.requestMethod = POST
+            connection.doOutput = true
 
             // Add default headers from the ZHttpClient
             client.getDefaultHeaders().forEach { (key, value) ->
@@ -57,6 +65,15 @@ class ZGet(val client: ZHttpClient) {
             // Add headers to the request
             headers?.forEach { (key, value) ->
                 connection.addRequestProperty(key, value)
+            }
+
+            // Serialize the request body to JSON
+            val body = client.getGsonInstance().toJson(requestBody)
+
+            // Write the request body to the output stream
+            connection.outputStream.use { outputStream ->
+                outputStream.write(body.toByteArray())
+                outputStream.flush()
             }
 
             val response = StringBuilder()
@@ -69,11 +86,11 @@ class ZGet(val client: ZHttpClient) {
                 }
             } catch (e: SocketTimeoutException) {
                 // If a socket timeout occurs, return an HttpResponse with the exception
-                Log.e(TAG, "doGet: $e", e)
+                Log.e(TAG, "doPost: $e", e)
                 return HttpResponse(exception = e)
             } catch (e: Exception) {
                 // If there's an error, read the error stream for additional information
-                Log.e(TAG, "doGet: $e", e)
+                Log.e(TAG, "doPost: $e", e)
                 BufferedReader(InputStreamReader(connection.errorStream)).use { reader ->
                     var line: String?
                     while (reader.readLine().also { line = it } != null) response.append(line)
@@ -86,11 +103,11 @@ class ZGet(val client: ZHttpClient) {
                 response.toString(),
                 connection.headerFields,
                 connection.date,
-                connection.permission,
+                connection.permission
             )
         } catch (e: Exception) {
             // If an exception occurs, log the error and return an HttpResponse with the exception
-            Log.e(TAG, "doGet: $e", e)
+            Log.e(TAG, "doPost: $e", e)
             HttpResponse(exception = e)
         } finally {
             // Disconnect the connection when done
@@ -99,22 +116,23 @@ class ZGet(val client: ZHttpClient) {
     }
 
     /**
-     * Performs a suspended GET HTTP request asynchronously, returning a [Deferred] object containing the result.
+     * Performs a suspended POST HTTP request asynchronously, returning a [Deferred] object containing the result.
      *
-     * @param endpoint The endpoint URL to send the GET request to.
+     * @param endpoint The endpoint URL to send the POST request to.
+     * @param requestBody The request body to include in the POST request.
      * @param queries The list of query parameters to include in the request.
      * @param headers The list of headers to include in the request.
-     * @return A [Deferred] object containing the result of the GET request.
+     * @return A [Deferred] object containing the result of the POST request.
      */
-    suspend fun doSuspendedGetRequest(
-        endpoint: String, queries: List<Query>?, headers: List<Header>?
+    suspend fun doSuspendedPostRequest(
+        endpoint: String, requestBody: Any, queries: List<Query>?, headers: List<Header>?
     ): Deferred<HttpResponse?> {
         return withContext(Dispatchers.IO) {
             async {
                 try {
-                    doGet(endpoint, queries, headers)
+                    doPost(endpoint, requestBody, queries, headers)
                 } catch (e: Exception) {
-                    Log.e(TAG, "doSuspendedGetRequest: $e", e)
+                    Log.e(TAG, "doSuspendedPostRequest: $e", e)
                     val response = HttpResponse(exception = e)
                     response
                 }
@@ -123,20 +141,22 @@ class ZGet(val client: ZHttpClient) {
     }
 
     /**
-     * Processes a GET HTTP request asynchronously.
+     * Processes a POST HTTP request asynchronously.
      *
-     * @param endpoint The endpoint URL to send the GET request to.
+     * @param endpoint The endpoint URL to send the POST request to.
+     * @param requestBody The request body to include in the POST request.
      * @param queries The list of query parameters to include in the request.
      * @param headers The list of headers to include in the request.
-     * @return A [Response] object containing the result of the GET request, or `null` if an error occurs.
+     * @return A [Response] object containing the result of the POST request, or `null` if an error occurs.
      */
-    suspend inline fun <reified T> processGet(
-        endpoint: String, queries: List<Query>?, headers: List<Header>?
+    suspend inline fun <reified T> processPost(
+        endpoint: String, requestBody: Any, queries: List<Query>?, headers: List<Header>?
     ): Response<T>? {
-        val response = doSuspendedGetRequest(endpoint, queries, headers).await() ?: return null
+        val response =
+            doSuspendedPostRequest(endpoint, requestBody, queries, headers).await() ?: return null
 
         response.exception?.let {
-            Log.e("ZGet", "processGet: $it", it)
+            Log.e("ZPost", "processPost: $it", it)
         }
 
         val body = client.getGsonInstance().deserializeBody<T>(response.body)
@@ -153,22 +173,24 @@ class ZGet(val client: ZHttpClient) {
     }
 
     /**
-     * Process a GET HTTP request with callback for the response.
+     * Process a POST HTTP request with callback for the response.
      *
      * @param endpoint Endpoint to append to the base URL.
-     * @param headers List of headers to include in the request.
+     * @param requestBody Body of the request.
      * @param queries List of query parameters to include in the URL.
+     * @param headers List of headers to include in the request.
      * @param onComplete Callback to handle the HttpResponse or an exception.
      * @return Job that will be completed with the HttpResponse or an exception.
      */
-    inline fun <reified T> processGet(
+    inline fun <reified T> processPost(
         endpoint: String,
+        requestBody: Any,
         queries: List<Query>?,
         headers: List<Header>?,
         crossinline onComplete: (success: Response<T>?, failure: Exception?) -> Unit
     ): Job {
         return CoroutineScope(Dispatchers.IO).launch {
-            val response = doSuspendedGetRequest(endpoint, queries, headers).await()
+            val response = doSuspendedPostRequest(endpoint, requestBody, queries, headers).await()
 
             if (response?.code == null) {
                 onComplete(null, NullPointerException("Could not make request."))
@@ -195,8 +217,7 @@ class ZGet(val client: ZHttpClient) {
     }
 
     companion object {
-        private const val TAG = "ZGet"
-        private const val GET = "GET"
+        private const val TAG = "ZPost"
+        private const val POST = "POST"
     }
-
 }
