+++ zhttp/src/main/java/com/zkrallah/zhttp/model/Basic.kt
@@ -1,4 +1,4 @@
-package com.zkrallah.zhttp
+package com.zkrallah.zhttp.model
 
 data class Basic(
     val username: String,
