+++ zhttp/src/main/java/com/zkrallah/zhttp/model/Bearer.kt
@@ -1,4 +1,4 @@
-package com.zkrallah.zhttp
+package com.zkrallah.zhttp.model
 
 /**
  * Represents a bearer token header.
