+++ zhttp/src/main/java/com/zkrallah/zhttp/model/HttpResponse.kt
@@ -1,4 +1,4 @@
-package com.zkrallah.zhttp
+package com.zkrallah.zhttp.model
 
 import java.lang.Exception
 import java.security.Permission
