+++ zhttp/src/main/java/com/zkrallah/zhttp/model/ZListener.kt
@@ -1,4 +1,4 @@
-package com.zkrallah.zhttp
+package com.zkrallah.zhttp.model
 
 /**
  * Interface for handling asynchronous HTTP requests.
